+++ core-ui/src/main/java/com/agileburo/anytype/core_ui/features/page/SupportNesting.kt
@@ -1,0 +1,6 @@
+package com.agileburo.anytype.core_ui.features.page
+
+/**
+ * Children of [BlockViewHolder] implementing this interface are expected to support nesting.
+ */
+interface SupportNesting